services/image-processing-service/get_emotion.py
import base64
import binascii # Import binascii for specific error handling
import cv2
import numpy as np
import random

# In a real scenario, you'd load a pre-trained model here
# For example:
# from deepface import DeepFace
# model = DeepFace.build_model("Emotion") 

def load_image_from_base64(base64_string: str):
    """Loads an image from a base64 string."""
    try:
        nparr = np.frombuffer(base64.b64decode(base64_string), np.uint8)
        img = cv2.imdecode(nparr, cv2.IMREAD_COLOR)
        return img
    except (binascii.Error, cv2.error, ValueError): # Catch base64 decoding errors or cv2 errors
        return None

def get_emotion(face_image_base64: str):
    """
    (Placeholder) Detects emotion from a cropped face image (base64 encoded).
    In a real implementation, this would use a pre-trained model.
    :param face_image_base64: Base64 encoded string of the cropped face image.
    :return: Predicted emotion (string) and confidence (float).
    """
    try:
        # Attempt to load image first to catch invalid base64
        face_image = load_image_from_base64(face_image_base64)
        if face_image is None:
            # If image loading fails, it's an invalid input for emotion detection
            raise ValueError("Could not load face image from base64 string.")

        # For demonstration, return a random emotion
        emotions = ["happy", "sad", "neutral", "angry", "surprise", "fear", "disgust"]
        predicted_emotion = random.choice(emotions)
        confidence = round(random.uniform(0.5, 0.99), 2) # Random confidence

        # In a real scenario:
        # result = DeepFace.analyze(face_image, actions=['emotion'], enforce_detection=False)
        # predicted_emotion = result[0]['dominant_emotion']
        # confidence = result[0]['emotion'][predicted_emotion] # Or other confidence metric

        return predicted_emotion, confidence
    except Exception as e:
        # Catch any errors during image loading or processing
        print(f"Error in emotion detection: {e}")
        return "unknown", 0.0

if __name__ == "__main__":
    # Example usage (requires a base64 encoded face image)
    print("This is a placeholder for emotion detection.")
    print("Please use it with a base64 encoded face image from crop_face.py output.")
    
    # Dummy base64 for testing (a very small black image)
    dummy_base64_image = "iVBORw0KGgoAAAANSUhEUgAAAAEAAAABCAQAAAC1HAwCAAAAC0lEQVR42mNkYAAAAAYAAjCB0C8AAAAASUVORK5CYII="
    emotion, conf = get_emotion(dummy_base64_image)
    print(f"Detected emotion: {emotion} with confidence: {conf}")
